Security, Services y Mappers terminados y documentados

+++ backend/src/main/java/com/viju/andaluciaskills/DTO/EvaluacionItemDTO.java
@@ -21,6 +21,6 @@ public class EvaluacionItemDTO {
     @JsonProperty("prueba_id_prueba")
     private Integer prueba_id_prueba;
 
-    private Integer valoracion;
+    private Double valoracion;
 
 }
